+++ apps/bot/src/commands/shitpost.slash.ts
@@ -1,4 +1,3 @@
-import { ArrayLoader } from "cocoa-discord-utils";
 import {
   CogSlashClass,
   Param,
@@ -6,11 +5,6 @@ import {
 } from "cocoa-discord-utils/slash/class";
 
 export class Shitpost extends CogSlashClass {
-  readonly kitaLoader = ArrayLoader.fromFile<string>(
-    "kitakita",
-    "data/kitakita.json",
-  );
-
   constructor() {
     super("Shitpost", "Cog for shitpost useless command");
   }
@@ -27,20 +21,5 @@ export class Shitpost extends CogSlashClass {
   ) {
     const who = gay ?? ctx.user;
     await ctx.reply(`<@${who.id}> is gay!`);
-  }
-
-  @SlashCommand(
-    "kitakita post from บจจิเดอะร็อคแฟนคลับไทยแลนด์ คิตะคิต้าโพสต์ติ้ง ✨⭐",
-  )
-  async kitakita(ctx: SlashCommand.Context) {
-    await this.kitaLoader.initialPromise;
-
-    const url = this.kitaLoader.getRandom();
-
-    if (!url) {
-      await ctx.reply("No kitakita post for unknown reason 😭");
-    } else {
-      await ctx.reply(`キタ～ン キタ～ン\n${url}`);
-    }
   }
 }