+++ apps/bot/patch.js
@@ -1,0 +1,3 @@
+import { patchPlayDl } from "@cocoa-discord/music-module";
+
+await patchPlayDl();