+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 Previous changelog before 4.0.605 may not be noted here.
 
+## [5.x.x] - 2023-??-??
+
+- refactor(bot): remove /lyrics (kashi module), /quality, /kitakita, /speak(TTS Module)
+
+- TODO: New website with better /quality
+
+- TODO: New TTS Module with Azure Cognitive API
+
 ## [4.1.710] - 2023-09-10
 
 - feat(goldenframe): get server avatar for goldenframe